refactor(EG-1040): seqera: refactor out EGRunPipelineParameterInputs
